Putting render-task to main.js
